LWCR-487: Consider keywords when sorting results by relevance.

+++ private-components/swingui/src/main/java/org/limewire/ui/swing/search/model/AbstractSimilarResultsDetector.java
@@ -87,10 +87,10 @@ public abstract class AbstractSimilarResultsDetector implements SimilarResultsDe
         VisualSearchResult parent2 = o2;
         VisualSearchResult parent3 = o1.getSimilarityParent();
         VisualSearchResult parent4 = o2.getSimilarityParent();
-        int parent1Count = parent1 == null ? 0 : parent1.getRelevance();
-        int parent2Count = parent2 == null ? 0 : parent2.getRelevance();
-        int parent3Count = parent3 == null ? 0 : parent3.getRelevance();
-        int parent4Count = parent4 == null ? 0 : parent4.getRelevance();
+        float parent1Count = parent1 == null ? 0 : parent1.getRelevance();
+        float parent2Count = parent2 == null ? 0 : parent2.getRelevance();
+        float parent3Count = parent3 == null ? 0 : parent3.getRelevance();
+        float parent4Count = parent4 == null ? 0 : parent4.getRelevance();
 
         if (parent4Count > parent3Count && parent4Count > parent2Count
                 && parent4Count > parent1Count) {
